backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
firebase-admin
python-multipart
pyrebase4
requests
SQLAlchemy
google-genai